Resources working

+++ database/migrations/2023_06_20_180552_create_user_game_table.php
@@ -16,9 +16,9 @@ class CreateUserGameTable extends Migration
         Schema::create('user_games', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('user_id');
-            $table->integer('metal')->default(50);
-            $table->integer('crystal')->default(50);
-            $table->integer('deuterium')->default(0);
+            $table->double('metal')->default(50);
+            $table->double('crystal')->default(50);
+            $table->double('deuterium')->default(0);
             $table->foreign('user_id')->references('id')->on('users')->cascadeOnDelete();
             $table->timestamps();
         });
